% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{assign_wells}
\alias{assign_wells}
\title{Add well locations to a volumetric experimental design}
\usage{
assign_wells(design, filename = "", platetype = "96-well", randomize = FALSE)
}
\arguments{
\item{design}{Volumetric experimental design}

\item{filename=""}{Optional output file. Default outputs to the console}

\item{platetype="96-well"}{Specifies the well naming scheme}

\item{randomize=FALSE}{If true, randomize the assigned wells. Else, put them in order.}
}
\description{
Add well locations to a volumetric experimental design
}
