% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{design_to_reagents}
\alias{design_to_reagents}
\title{Convert a volume based experimental design to an ordered reagent dataframe}
\usage{
design_to_reagents(design_df, plate)
}
\arguments{
\item{design_df}{Formatted dataframe with columns for each reagent, followed by either well location or row col}

\item{plate}{plate information}
}
\description{
Convert a volume based experimental design to an ordered reagent dataframe
}
