% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{dispense_list}
\alias{dispense_list}
\title{Main running funciton. Turn an experimental design into a dispense list}
\usage{
dispense_list(design, name, platetype = "breakaway_pcr_96", ...)
}
\arguments{
\item{design}{an Nx(M+2) dataframe containing N experiment rows and M reagent columns. Each entry is a volume in µL. Two extra columns called "Row" (ex. "A") and "Col" (ex. "1") must be included to denote well positions.}

\item{name}{name of the experiment. Appropriate extension is added automatically.}

\item{platetype="breakaway_pcr_96"}{specifies the plate type. Must be one of: "96-well", "breakaway_pcr_96" , 384-well".}
}
\description{
Main running funciton. Turn an experimental design into a dispense list
}
