% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{export_mantis_worklist}
\alias{export_mantis_worklist}
\title{Convert a reagent plate layout to a mantis dispense list}
\usage{
export_mantis_worklist(filename, reagent_layout)
}
\arguments{
\item{filename}{a filename for the dispense list. must end in ".dl.txt"}

\item{reagent_layout}{a list of reagent volume matricies. see 'reagents_to_layout'}
}
\description{
Convert a reagent plate layout to a mantis dispense list
}
