% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateLayout.R
\name{install_plate_layout}
\alias{install_plate_layout}
\title{Install the julia pacakges for plate layout if needed}
\usage{
install_plate_layout()
}
\description{
Install the julia pacakges for plate layout if needed
}
