% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Normalize.R
\name{mantis_normalize}
\alias{mantis_normalize}
\title{Normalize DNA concentrations with water using the Mantis}
\usage{
mantis_normalize(
  input_path,
  platetype = "breakaway_pcr_96",
  maxvol = 200,
  usenanomolar = FALSE
)
}
\arguments{
\item{input_path}{path to the normalization input template. Template can be found at https://github.com/bmdavid2/DispenseFormulatrix/blob/main/mantis_normalize_template.csv}

\item{platetype="breakaway_pcr_96"}{specifies the plate type. Must be one of: "96-well", "breakaway_pcr_96" , 384-well".}

\item{maxvol=200}{the maximum volume allowed for the final concentration}

\item{usenanomolar=FALSE}{standard final concentrations are ng/ul}
}
\description{
Normalize DNA concentrations with water using the Mantis
}
