% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{multi_dispense}
\alias{multi_dispense}
\title{Make multiple dispense lists for a single experiment. Turn an experimental design into a dispense list}
\usage{
multi_dispense(designs, name, platetype = "breakaway_pcr_96")
}
\arguments{
\item{name}{name of the experiment. Appropriate extensions are added automatically.}

\item{design}{a list of  Nx(M+2) dataframes containing N experiment rows and M reagent columns. Each entry is a volume in µL. Two extra columns called "Row" (ex. "A") and "Col" (ex. "1") must be included to denote well positions.}

\item{platetype="breakaway_pcr_96"}{specifies the plate type. Must be one of: "96-well", "breakaway_pcr_96" , 384-well".}
}
\description{
Make multiple dispense lists for a single experiment. Turn an experimental design into a dispense list
}
