% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateLayout.R
\name{plate_layout}
\alias{plate_layout}
\title{Automatically create an experimental plate layout that minimizes pipetting operations}
\usage{
plate_layout(design, factors, types, num_generations = 50, popsize = 30)
}
\arguments{
\item{design}{an experimental design dataframe containing the coded level of each factor}

\item{factors}{an array of the factor names}

\item{types}{an array of the factor types: must be "auto" ,"col" "wafer","plate}

\item{num_generations}{the number of GA generations}

\item{popsize}{the GA population size}
}
\description{
Automatically create an experimental plate layout that minimizes pipetting operations
}
