% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{plateinfo}
\alias{plateinfo}
\title{Gather and Store all relevant information about the plate being used}
\usage{
plateinfo(platetype = "breakaway_pcr_96")
}
\arguments{
\item{platetype="breakaway_pcr_96"}{specifies the plate type. Must be one of: "96-well", "breakaway_pcr_96" , 384-well".}
}
\description{
Gather and Store all relevant information about the plate being used
}
