% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Design.R
\name{process_design}
\alias{process_design}
\title{Convert a coded experimental design into a volume-based design for automated dispensing}
\usage{
process_design(
  coded_design,
  reagent_names,
  volume_levels,
  separate_stocks,
  dispense_group
)
}
\arguments{
\item{coded_design}{}
}
\description{
Convert a coded experimental design into a volume-based design for automated dispensing
}
