% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{reagents_to_layout}
\alias{reagents_to_layout}
\title{Convert an ordered reagent volume dataframe to a plate layout}
\usage{
reagents_to_layout(reagent_df, platetype = "96-well")
}
\arguments{
\item{reagent_df}{reagent volumes in the order they should appear on the plate. Must only be reagent volumes}

\item{platetype="96-well"}{specifies the plate type for the layout}
}
\description{
Convert an ordered reagent volume dataframe to a plate layout
}
