% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Design.R
\name{stock_volumes}
\alias{stock_volumes}
\title{Get the stock volume required for each reagent}
\usage{
stock_volumes(processed_design, priming_buffer = 10, uncertainty = 0.1)
}
\arguments{
\item{processed_design}{}
}
\description{
Get the stock volume required for each reagent
}
