% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{assign_wells}
\alias{assign_wells}
\title{Add well locations to a volumetric experimental design}
\usage{
assign_wells(
  design,
  filename = "",
  platetype = "breakaway_pcr_96",
  randomize = FALSE
)
}
\arguments{
\item{design}{Volumetric experimental design}

\item{filename=""}{Optional output file. Default outputs to the console}

\item{platetype="breakaway_pcr_96"}{Must be either "96-well" or "384-well"}

\item{randomize=FALSE}{If true, randomize the assigned wells. Else, put them in order.}
}
\description{
Add well locations to a volumetric experimental design
}
