% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{design_to_reagents}
\alias{design_to_reagents}
\title{Convert a volume based experimental design to an ordered reagent dataframe}
\usage{
design_to_reagents(design_df, platetype = "96-well", Well = TRUE)
}
\arguments{
\item{design_df}{Formatted dataframe with columns for each reagent, followed by either well location or row col}

\item{platetype="96-well"}{specifies the plate type for the design}

\item{Well=TRUE}{if Well=TRUE, then use well designation, else combine row and col into well info}
}
\description{
Convert a volume based experimental design to an ordered reagent dataframe
}
