% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantis_Dispense.R
\name{make_dispense_list}
\alias{make_dispense_list}
\title{Main running funciton. Turn an experimental design into a dispense list}
\usage{
make_dispense_list(design, name, platetype = "breakaway_pcr_96", ...)
}
\arguments{
\item{design}{Volumetric experimental design. Must include either Well or Row & Col columns}

\item{name}{Design name. Funciton adds appropriate suffix automatically}

\item{platetype="96-well"}{specifies the plate type. Must be one of: \n "96-well" \n "breakaway_pcr_96" \n 384-well".}
}
\description{
Main running funciton. Turn an experimental design into a dispense list
}
